+++ R/lockfile.R
@@ -132,13 +132,19 @@ renv_lockfile_sort <- function(lockfile) {
 
 }
 
-renv_lockfile_create <- function(project, library, type) {
+renv_lockfile_create <- function(project, library, type, packages) {
 
   lockfile <- renv_lockfile_init(project)
 
   renv_records(lockfile) <-
-    renv_snapshot_r_packages(library = library, project = project) %>%
-    renv_snapshot_filter(project = project, type = type) %>%
+
+    renv_snapshot_r_packages(library = library,
+                             project = project) %>%
+
+    renv_snapshot_filter(project = project,
+                         type = type,
+                         packages = packages) %>%
+
     renv_snapshot_fixup()
 
   lockfile <- renv_lockfile_fini(lockfile)